fix: move types to the types folder
