+++ stonecrop/src/stonecrop.ts
@@ -14,22 +14,7 @@ import { createHST, type HSTNode } from './stores/hst'
 import { useOperationLogStore } from './stores/operation-log'
 import type { OperationLogConfig } from './types/operation-log'
 import type { RouteContext } from './types/registry'
-
-/**
- * Options for constructing a Stonecrop instance directly.
- * When using the Vue plugin, pass these via `InstallOptions` instead.
- * @public
- */
-export interface StonecropOptions {
-	/**
-	 * Data client for fetching doctype metadata and records.
-	 * Use \@stonecrop/graphql-client's StonecropClient for GraphQL backends,
-	 * or implement DataClient for custom data sources.
-	 *
-	 * Can be set later via `setClient()` for deferred configuration.
-	 */
-	client?: DataClient
-}
+import type { StonecropOptions } from './types/stonecrop'
 
 /**
  * Main Stonecrop class with HST integration and built-in Operation Log
